+++ packages/jest-circus/src/legacy-code-todo-rewrite/jestAdapterInit.ts
@@ -51,8 +51,8 @@ export const initialize = async ({
   config: Config.ProjectConfig;
   environment: JestEnvironment;
   globalConfig: Config.GlobalConfig;
-  localRequire: <T = unknown>(path: Config.Path) => T;
-  testPath: Config.Path;
+  localRequire: <T = unknown>(path: string) => T;
+  testPath: string;
   parentProcess: NodeJS.Process;
   sendMessageToJest?: TestFileEvent;
   setGlobalsForRuntime: (globals: RuntimeGlobals) => void;