+++ packages/jest-types/src/Config.ts
@@ -14,10 +14,6 @@ type CoverageProvider = 'babel' | 'v8';
 
 type Timers = 'real' | 'fake' | 'modern' | 'legacy';
 
-export type Path = string;
-
-export type Glob = string;
-
 export type HasteConfig = {
   /** Whether to hash files using SHA-1. */
   computeSha1?: boolean;
@@ -31,7 +27,7 @@ export type HasteConfig = {
    *   Projects with `watchman` set to true will error if this option is set to true.
    */
   enableSymlinks?: boolean;
-  /** Path to a custom implementation of Haste. */
+  /** string to a custom implementation of Haste. */
   hasteImplModulePath?: string;
   /** All platforms to target, e.g ['ios', 'android']. */
   platforms?: Array<string>;
@@ -65,7 +61,7 @@ export type DefaultOptions = {
   automock: boolean;
   bail: number;
   cache: boolean;
-  cacheDirectory: Path;
+  cacheDirectory: string;
   changedFilesWithAncestor: boolean;
   ci: boolean;
   clearMocks: boolean;
@@ -77,8 +73,8 @@ export type DefaultOptions = {
   detectOpenHandles: boolean;
   errorOnDeprecated: boolean;
   expand: boolean;
-  extensionsToTreatAsEsm: Array<Path>;
-  forceCoverageMatch: Array<Glob>;
+  extensionsToTreatAsEsm: Array<string>;
+  forceCoverageMatch: Array<string>;
   globals: ConfigGlobals;
   haste: HasteConfig;
   injectGlobals: boolean;
@@ -97,26 +93,26 @@ export type DefaultOptions = {
   resetMocks: boolean;
   resetModules: boolean;
   restoreMocks: boolean;
-  roots: Array<Path>;
+  roots: Array<string>;
   runTestsByPath: boolean;
   runner: string;
-  setupFiles: Array<Path>;
-  setupFilesAfterEnv: Array<Path>;
+  setupFiles: Array<string>;
+  setupFilesAfterEnv: Array<string>;
   skipFilter: boolean;
   slowTestThreshold: number;
-  snapshotSerializers: Array<Path>;
+  snapshotSerializers: Array<string>;
   testEnvironment: string;
   testEnvironmentOptions: Record<string, unknown>;
   testFailureExitCode: string | number;
   testLocationInResults: boolean;
-  testMatch: Array<Glob>;
+  testMatch: Array<string>;
   testPathIgnorePatterns: Array<string>;
   testRegex: Array<string>;
   testRunner: string;
   testSequencer: string;
   testURL: string;
   timers: Timers;
-  transformIgnorePatterns: Array<Glob>;
+  transformIgnorePatterns: Array<string>;
   useStderr: boolean;
   watch: boolean;
   watchPathIgnorePatterns: Array<string>;
@@ -140,13 +136,13 @@ export type InitialOptions = Partial<{
   automock: boolean;
   bail: boolean | number;
   cache: boolean;
-  cacheDirectory: Path;
+  cacheDirectory: string;
   ci: boolean;
   clearMocks: boolean;
   changedFilesWithAncestor: boolean;
   changedSince: string;
   collectCoverage: boolean;
-  collectCoverageFrom: Array<Glob>;
+  collectCoverageFrom: Array<string>;
   collectCoverageOnlyFrom: {
     [key: string]: boolean;
   };
@@ -160,11 +156,11 @@ export type InitialOptions = Partial<{
   detectOpenHandles: boolean;
   displayName: string | DisplayName;
   expand: boolean;
-  extensionsToTreatAsEsm: Array<Path>;
+  extensionsToTreatAsEsm: Array<string>;
   extraGlobals: Array<string>;
-  filter: Path;
+  filter: string;
   findRelatedTests: boolean;
-  forceCoverageMatch: Array<Glob>;
+  forceCoverageMatch: Array<string>;
   forceExit: boolean;
   json: boolean;
   globals: ConfigGlobals;
@@ -180,7 +176,7 @@ export type InitialOptions = Partial<{
   maxWorkers: number | string;
   moduleDirectories: Array<string>;
   moduleFileExtensions: Array<string>;
-  moduleLoader: Path;
+  moduleLoader: string;
   moduleNameMapper: {
     [key: string]: string | Array<string>;
   };
@@ -192,52 +188,52 @@ export type InitialOptions = Partial<{
   notifyMode: string;
   onlyChanged: boolean;
   onlyFailures: boolean;
-  outputFile: Path;
+  outputFile: string;
   passWithNoTests: boolean;
   /**
    * @deprecated Use `transformIgnorePatterns` options instead.
    */
-  preprocessorIgnorePatterns: Array<Glob>;
+  preprocessorIgnorePatterns: Array<string>;
   preset: string | null | undefined;
   prettierPath: string | null | undefined;
-  projects: Array<Glob | InitialProjectOptions>;
+  projects: Array<string | InitialProjectOptions>;
   replname: string | null | undefined;
   resetMocks: boolean;
   resetModules: boolean;
-  resolver: Path | null | undefined;
+  resolver: string | null | undefined;
   restoreMocks: boolean;
-  rootDir: Path;
-  roots: Array<Path>;
+  rootDir: string;
+  roots: Array<string>;
   runner: string;
   runTestsByPath: boolean;
   /**
    * @deprecated Use `transform` options instead.
    */
   scriptPreprocessor: string;
-  setupFiles: Array<Path>;
+  setupFiles: Array<string>;
   /**
    * @deprecated Use `setupFilesAfterEnv` options instead.
    */
-  setupTestFrameworkScriptFile: Path;
-  setupFilesAfterEnv: Array<Path>;
+  setupTestFrameworkScriptFile: string;
+  setupFilesAfterEnv: Array<string>;
   silent: boolean;
   skipFilter: boolean;
   skipNodeResolution: boolean;
   slowTestThreshold: number;
-  snapshotResolver: Path;
-  snapshotSerializers: Array<Path>;
+  snapshotResolver: string;
+  snapshotSerializers: Array<string>;
   snapshotFormat: SnapshotFormat;
   errorOnDeprecated: boolean;
   testEnvironment: string;
   testEnvironmentOptions: Record<string, unknown>;
   testFailureExitCode: string | number;
   testLocationInResults: boolean;
-  testMatch: Array<Glob>;
+  testMatch: Array<string>;
   testNamePattern: string;
   /**
    * @deprecated Use `roots` options instead.
    */
-  testPathDirs: Array<Path>;
+  testPathDirs: Array<string>;
   testPathIgnorePatterns: Array<string>;
   testRegex: string | Array<string>;
   testResultsProcessor: string;
@@ -247,9 +243,9 @@ export type InitialOptions = Partial<{
   testTimeout: number;
   timers: Timers;
   transform: {
-    [regex: string]: Path | TransformerConfig;
+    [regex: string]: string | TransformerConfig;
   };
-  transformIgnorePatterns: Array<Glob>;
+  transformIgnorePatterns: Array<string>;
   watchPathIgnorePatterns: Array<string>;
   unmockedModulePathPatterns: Array<string>;
   updateSnapshot: boolean;
@@ -289,7 +285,7 @@ export type GlobalConfig = {
   changedFilesWithAncestor: boolean;
   ci: boolean;
   collectCoverage: boolean;
-  collectCoverageFrom: Array<Glob>;
+  collectCoverageFrom: Array<string>;
   collectCoverageOnlyFrom?: {
     [key: string]: boolean;
   };
@@ -301,7 +297,7 @@ export type GlobalConfig = {
   detectLeaks: boolean;
   detectOpenHandles: boolean;
   expand: boolean;
-  filter?: Path;
+  filter?: string;
   findRelatedTests: boolean;
   forceExit: boolean;
   json: boolean;
@@ -317,15 +313,15 @@ export type GlobalConfig = {
   noSCM?: boolean;
   notify: boolean;
   notifyMode: NotifyMode;
-  outputFile?: Path;
+  outputFile?: string;
   onlyChanged: boolean;
   onlyFailures: boolean;
   passWithNoTests: boolean;
-  projects: Array<Glob>;
+  projects: Array<string>;
   replname?: string;
   reporters?: Array<string | ReporterConfig>;
   runTestsByPath: boolean;
-  rootDir: Path;
+  rootDir: string;
   silent?: boolean;
   skipFilter: boolean;
   snapshotFormat: SnapshotFormat;
@@ -351,19 +347,19 @@ export type GlobalConfig = {
 export type ProjectConfig = {
   automock: boolean;
   cache: boolean;
-  cacheDirectory: Path;
+  cacheDirectory: string;
   clearMocks: boolean;
   coveragePathIgnorePatterns: Array<string>;
-  cwd: Path;
+  cwd: string;
   dependencyExtractor?: string;
   detectLeaks: boolean;
   detectOpenHandles: boolean;
   displayName?: DisplayName;
   errorOnDeprecated: boolean;
-  extensionsToTreatAsEsm: Array<Path>;
+  extensionsToTreatAsEsm: Array<string>;
   extraGlobals: Array<keyof typeof globalThis>;
-  filter?: Path;
-  forceCoverageMatch: Array<Glob>;
+  filter?: string;
+  forceCoverageMatch: Array<string>;
   globalSetup?: string;
   globalTeardown?: string;
   globals: ConfigGlobals;
@@ -371,7 +367,7 @@ export type ProjectConfig = {
   injectGlobals: boolean;
   moduleDirectories: Array<string>;
   moduleFileExtensions: Array<string>;
-  moduleLoader?: Path;
+  moduleLoader?: string;
   moduleNameMapper: Array<[string, string]>;
   modulePathIgnorePatterns: Array<string>;
   modulePaths?: Array<string>;
@@ -379,30 +375,30 @@ export type ProjectConfig = {
   prettierPath: string;
   resetMocks: boolean;
   resetModules: boolean;
-  resolver?: Path;
+  resolver?: string;
   restoreMocks: boolean;
-  rootDir: Path;
-  roots: Array<Path>;
+  rootDir: string;
+  roots: Array<string>;
   runner: string;
-  setupFiles: Array<Path>;
-  setupFilesAfterEnv: Array<Path>;
+  setupFiles: Array<string>;
+  setupFilesAfterEnv: Array<string>;
   skipFilter: boolean;
   skipNodeResolution?: boolean;
   slowTestThreshold: number;
-  snapshotResolver?: Path;
-  snapshotSerializers: Array<Path>;
+  snapshotResolver?: string;
+  snapshotSerializers: Array<string>;
   snapshotFormat: SnapshotFormat;
   testEnvironment: string;
   testEnvironmentOptions: Record<string, unknown>;
-  testMatch: Array<Glob>;
+  testMatch: Array<string>;
   testLocationInResults: boolean;
   testPathIgnorePatterns: Array<string>;
   testRegex: Array<string | RegExp>;
   testRunner: string;
   testURL: string;
   timers: Timers;
-  transform: Array<[string, Path, Record<string, unknown>]>;
-  transformIgnorePatterns: Array<Glob>;
+  transform: Array<[string, string, Record<string, unknown>]>;
+  transformIgnorePatterns: Array<string>;
   watchPathIgnorePatterns: Array<string>;
   unmockedModulePathPatterns?: Array<string>;
 };